Replace babelmarkfix=yes everywhere with babelmarkfix=unchanged
\documentclass{book}
\usepackage[T1]{fontenc}
\PassOptionsToPackage{defaults=hu-min}{magyar.ldf}
\usepackage[magyar]{babel}
\begin{document}
~\newpage~
\end{document}

In the generated PDF, the headers contain test text ("BB" and "BA").
The reason is the babelmarkfix=yes option.
The problem is caused by the new mark mechanism of the kernel.
This option is not worth fixing, because it is no longer needed (deprecated). It will be deleted later. We just turned it off everywhere now.