Fixed the option mathbrk=fix
\DocumentMetadata{lang=hu-HU}
\documentclass{book}
\usepackage[T1]{fontenc}
\PassOptionsToPackage{defaults=hu-min}{hungarian.ldf}
\usepackage{babel}
\begin{document}
\title{Cím}
\author{Név}
\maketitle
\tableofcontents
\chapter{Cím}
Szöveg
\end{document}

An error occurs when generating the pdf.
The reason is the mathbrk=fix option, which is called by defaults=hu-min.
LaTeX-lab (PDF-tagging/math-grabbing) introduced the \UseMathForPositioningText for developers, which must be followed directly by a $ sign.
The mathbrk=fix option redefine the beginning of \@tabular and inserts a \def before the very first $ to eliminate the effect of \@@magyar@domathbins.
Thus, mathbrk=fix and \UseMathForPositioningText conflict.